feat, refactor: 차트 부분 리팩토링 => 컴포넌트 나누기 state값 위에서 Context API로 관리하도록 작업
=> 컴포넌트를 나누고 state값을 조정하는 부분에서 예기치못한 버그 발생.. 해당 현상 수정 후 바로 PR 보낼게요 우선 merge할게요!
